+++ rocketpy/rocket/aero_surface/fins/fins.py
@@ -380,10 +380,26 @@ class Fins(AeroSurface):
 
         Parameters
         ----------
-        stream_speed : int, float
-            Speed of the flow stream in the body frame.
+        stream_velocity : tuple of float
+            The velocity of the airflow relative to the surface.
+        stream_speed : float
+            The magnitude of the airflow speed.
+        stream_mach : float
+            The Mach number of the airflow.
+        rho : float
+            Air density.
+        cp : Vector
+            Center of pressure coordinates in the body frame.
+        omega: tuple[float, float, float]
+            Tuple containing angular velocities around the x, y, z axes.
 
+        Returns
+        -------
+        tuple of float
+            The aerodynamic forces (lift, side_force, drag) and moments
+            (pitch, yaw, roll) in the body frame.
         """
+
         R1, R2, R3, M1, M2, _ = super().compute_forces_and_moments(
             stream_velocity,
             stream_speed,